feat(ExposedPortSetting): add port strategy support to GenericContainer
